% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.bouldr.R
\name{plot.bouldr}
\alias{plot.bouldr}
\title{Plot the ROC objects in a pretty way, with faceting when appropriate}
\usage{
\method{plot}{bouldr}(x, ...)
}
\arguments{
\item{x}{An object of class \code{bouldr}}

\item{...}{Additional arguments}
}
\value{
A ggplot2 object with relevant plot(s)
}
\description{
Plot the ROC objects in a pretty way, with faceting when appropriate
}
