% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bouldr.R
\name{print.bouldr}
\alias{print.bouldr}
\title{A print function for the bouldr object}
\usage{
\method{print}{bouldr}(x, tests = FALSE, ...)
}
\arguments{
\item{x}{A bouldr object to print}

\item{tests}{Should the tests be printed? Default is FALSE, since these can be long}

\item{...}{Additional arguments for printing}
}
\description{
Prints the original formula, then the AUCs then (optionally) the tests
}
