% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tumble_rocs.bouldr.R
\name{tumble_rocs}
\alias{tumble_rocs}
\title{An internal function for extracing data from bouldr objects for plotting}
\usage{
tumble_rocs(rocbag)
}
\arguments{
\item{rocbag}{The bouldr object to manipulate}
}
\value{
A data frame with plottable data
}
\description{
An internal function for extracing data from bouldr objects for plotting
}
